Struct embassy_usb::class::hid::HidReaderWriter

pub struct HidReaderWriter<'d, D: Driver<'d>, const READ_N: usize, const WRITE_N: usize> { /* private fields */ }

USB HID reader/writer.

Implementations

impl<'d, D: Driver<'d>, const READ_N: usize, const WRITE_N: usize> HidReaderWriter<'d, D, READ_N, WRITE_N>

pub fn new( builder: &mut Builder<'d, D>, state: &'d mut State<'d>, config: Config<'d> ) -> Self

Creates a new HidReaderWriter.

This will allocate one IN and one OUT endpoints. If you only need writing (sending) HID reports, consider using HidWriter::new instead, which allocates an IN endpoint only.

pub fn split(self) -> (HidReader<'d, D, READ_N>, HidWriter<'d, D, WRITE_N>)

Splits into separate readers/writers for input and output reports.

pub async fn ready(&mut self)

Waits for both IN and OUT endpoints to be enabled.

pub async fn write_serialize<IR: AsInputReport>( &mut self, r: &IR ) -> Result<(), EndpointError>

Writes an input report by serializing the given report structure.

pub async fn write(&mut self, report: &[u8]) -> Result<(), EndpointError>

Writes report to its interrupt endpoint.

pub async fn read(&mut self, buf: &mut [u8]) -> Result<usize, ReadError>

Reads an output report from the Interrupt Out pipe.

See HidReader::read.