Struct embassy_usb::Builder

pub struct Builder<'d, D: Driver<'d>> { /* private fields */ }

UsbDevice builder.

Implementations

impl<'d, D: Driver<'d>> Builder<'d, D>

pub fn new( driver: D, config: Config<'d>, config_descriptor_buf: &'d mut [u8], bos_descriptor_buf: &'d mut [u8], msos_descriptor_buf: &'d mut [u8], control_buf: &'d mut [u8] ) -> Self

Creates a builder for constructing a new UsbDevice.

control_buf is a buffer used for USB control request data. It should be sized large enough for the length of the largest control request (in or out) anticipated by any class added to the device.

pub fn build(self) -> UsbDevice<'d, D>

Creates the UsbDevice instance with the configuration in this builder.

pub fn control_buf_len(&self) -> usize

Returns the size of the control request data buffer. Can be used by classes to validate the buffer is large enough for their needs.

pub fn function( &mut self, class: u8, subclass: u8, protocol: u8 ) -> FunctionBuilder<'_, 'd, D>

Add an USB function.

If Config::composite_with_iads is set, this will add an IAD descriptor with the given class/subclass/protocol, associating all the child interfaces.

If it’s not set, no IAD descriptor is added.

pub fn handler(&mut self, handler: &'d mut dyn Handler)

Add a Handler.

The Handler is called on some USB bus events, and to handle all control requests not already handled by the USB stack.

pub fn string(&mut self) -> StringIndex

Allocates a new string index.

pub fn msos_descriptor(&mut self, windows_version: u32, vendor_code: u8)

Add an MS OS 2.0 Descriptor Set.

Panics if called more than once.

pub fn msos_feature<T: DeviceLevelDescriptor>(&mut self, desc: T)

Add an MS OS 2.0 Device Level Feature Descriptor.

pub fn msos_writer(&mut self) -> &mut MsOsDescriptorWriter<'d>

Gets the underlying MsOsDescriptorWriter to allow adding subsets and features for classes that do not add their own.